requirements.txt
streamlit
openai
faiss-cpu
sentence-transformers
transformers
torch
PyMuPDF
markdown
pytest
